wip: one singular autosave, handle save slot ui more reactively (no full clear, full init, on change)

+++ Yolk.ExampleGame/ui/game_save_slot/GameSaveSlot.cs
@@ -3,6 +3,7 @@ namespace Yolk.UI;
 using Chickensoft.AutoInject;
 using Chickensoft.Introspection;
 using Godot;
+using Yolk.Data;
 using Yolk.FS;
 using Yolk.Game;
 
@@ -10,8 +11,19 @@ using Yolk.Game;
 public partial class GameSaveSlot : PanelContainer {
   public override void _Notification(int what) => this.Notify(what);
 
+  private IGodotSaveInfo? _saveInfo;
+
   public bool AllowSave { get; set; } = true;
-  public IGodotSaveInfo SaveInfo { get; set; } = default!;
+
+  public IGodotSaveInfo? SaveInfo {
+    get => _saveInfo;
+    set {
+      _saveInfo = value;
+      if (IsInsideTree()) {
+        UpdateVisuals();
+      }
+    }
+  }
 
   [Dependency] private IGameRepo GameRepo => this.DependOn<IGameRepo>();
 
@@ -24,11 +36,31 @@ public partial class GameSaveSlot : PanelContainer {
   public void OnResolved() {
     SaveButton.Visible = AllowSave;
 
-    PreviewImage.Texture = GodotSaver.GetThumbnail(SaveInfo.SaveName);
+    if (SaveInfo is null) {
+      GD.PushWarning("SaveInfo is null, cannot initialize GameSaveSlot.");
+      return;
+    }
 
-    SaveNameLabel.Text = SaveInfo.SaveName;
+    UpdateVisuals();
 
+    SaveButton.Pressed += () => GameRepo.Save(SaveInfo.SaveName);
     LoadButton.Pressed += () => GameRepo.Load(SaveInfo.SaveName);
     DeleteButton.Pressed += () => GameRepo.Delete(SaveInfo.SaveName);
+  }
+
+  public void Update(string saveName) {
+    SaveInfo = GodotSaver.GetSaveInfo<GameData>(saveName);
+
+    UpdateVisuals();
+  }
+
+  public void UpdateVisuals() {
+    if (SaveInfo is null) {
+      GD.PrintErr($"No save info for node {Name}.");
+      return;
+    }
+
+    PreviewImage.Texture = GodotSaver.GetThumbnail(SaveInfo.SaveName);
+    SaveNameLabel.Text = SaveInfo.SaveName;
   }
 }
